+++ lib/generators/cms/component/blog/templates/app/controllers/blog_entry_controller.rb
@@ -1,8 +1,4 @@
 class BlogEntryController < CmsController
-  layout 'blog'
-
   def index
-    @entry = @obj
-    @blog = @entry.blog
   end
 end